banner_shield is now vanilla_shield for consistency.

+++ common/src/main/java/com/github/stellarwind22/shieldlib/init/ShieldLibClient.java
@@ -70,7 +70,7 @@ public class ShieldLibClient {
                     BlocksAttacks blocksAttacks = stack.get(DataComponents.BLOCKS_ATTACKS);
 
                     if(blocksAttacks != null) {
-                        float cooldownTicks = ShieldLib.getCooldownTicksWithModifiers(player, stack, blocksAttacks);
+                        double cooldownTicks = ShieldLib.getCooldownTicksWithModifiers(player, stack, blocksAttacks);
 
                         tooltip.add(Component.literal(""));
                         tooltip.add(Component.translatable("shieldlib.shield_tooltip.start")
